requirements.txt
Django>=5.0
psycopg2-binary>=2.9
PyJWT>=2.0
